+++ packages/cli/src/commands/integration/command.ts
@@ -1,30 +1,15 @@
 import { packageName } from '../../util/pkg-name';
-import { Command } from '../help';
 
-export const integrationCommand: Command = {
-  name: 'integration',
-  description: 'Manage marketplace integrations',
-  options: [],
+export const addSubCommand = {
+  name: 'add',
+  description: 'Installs a marketplace integration',
   arguments: [
     {
-      name: 'command',
+      name: 'name',
       required: true,
     },
   ],
-  subcommands: [
-    {
-      name: 'add',
-      description: 'Installs a marketplace integration',
-      arguments: [
-        {
-          name: 'name',
-          required: true,
-        },
-      ],
-      options: [],
-      examples: [],
-    },
-  ],
+  options: [],
   examples: [
     {
       name: 'Install a marketplace integration',
@@ -34,4 +19,88 @@ export const integrationCommand: Command = {
       ],
     },
   ],
-};
+} as const;
+
+export const openSubCommand = {
+  name: 'open',
+  description: "Opens a marketplace integration's dashboard",
+  arguments: [
+    {
+      name: 'name',
+      required: true,
+    },
+  ],
+  options: [],
+  examples: [
+    {
+      name: "Open a marketplace integration's dashboard",
+      value: [
+        `${packageName} integration open <integration-name>`,
+        `${packageName} integration open acme`,
+      ],
+    },
+  ],
+} as const;
+
+export const listSubcommand = {
+  name: 'list',
+  description: 'Lists all resources from marketplace integrations',
+  arguments: [
+    {
+      name: 'project',
+      required: false,
+    },
+  ],
+  options: [
+    {
+      name: 'integration',
+      description: 'limits the resources listed to a designated integration',
+      shorthand: 'i',
+      type: String,
+      deprecated: false,
+      argument: 'NAME',
+    },
+    {
+      name: 'all',
+      description: 'lists all resources regardless of project',
+      shorthand: 'a',
+      type: Boolean,
+      deprecated: false,
+    },
+  ],
+  examples: [
+    {
+      name: 'List all resources',
+      value: [`${packageName} integrations list`],
+    },
+    {
+      name: 'Filter the resources to a single integration',
+      value: [
+        `${packageName} integrations list --integration <integration>`,
+        `${packageName} integrations list --integration acme`,
+        `${packageName} integrations list -i acme`,
+      ],
+    },
+    {
+      name: 'List all marketplace resources for the current team',
+      value: [
+        `${packageName} integrations list --all`,
+        `${packageName} integrations list -a`,
+      ],
+    },
+  ],
+} as const;
+
+export const integrationCommand = {
+  name: 'integration',
+  description: 'Manage marketplace integrations',
+  options: [],
+  arguments: [
+    {
+      name: 'command',
+      required: true,
+    },
+  ],
+  subcommands: [addSubCommand, openSubCommand, listSubcommand],
+  examples: [],
+} as const;