+++ packages/cli/src/commands/integration/types.ts
@@ -37,6 +37,21 @@ export interface IntegrationProduct {
   metadataSchema: MetadataSchema;
 }
 
+export type InstallationType = 'marketplace' | 'external';
+
+export interface Configuration {
+  id: string;
+  integrationId: string;
+  ownerId: string;
+  slug: string;
+  teamId: string;
+  userId: string;
+  scopes: string[];
+  source: string;
+  installationType: InstallationType;
+  projects: string[];
+}
+
 export interface Integration {
   id: string;
   slug: string;
@@ -46,7 +61,7 @@ export interface Integration {
 
 export interface IntegrationInstallation {
   id: string;
-  installationType: 'marketplace' | 'external';
+  installationType: InstallationType;
   ownerId: string;
 }
 
@@ -67,4 +82,22 @@ export interface BillingPlan {
     value?: string;
   }[];
   disabled?: boolean;
+}
+
+export interface Resource {
+  id: string;
+  type: string;
+  name: string;
+  status?: string | null;
+  product?: {
+    name?: string;
+    slug?: string;
+    integrationConfigurationId?: string;
+  };
+  projectsMetadata?: {
+    id: string;
+    projectId: string;
+    name: string;
+    environments: string[];
+  }[];
 }