+++ packages/cli/src/index.ts
@@ -554,6 +554,7 @@ const main = async () => {
       let func: any;
       switch (targetCommand) {
         case 'alias':
+          telemetry.trackCliCommandAlias(userSuppliedSubCommand);
           func = require('./commands/alias').default;
           break;
         case 'bisect':
@@ -566,6 +567,7 @@ const main = async () => {
           func = require('./commands/certs').default;
           break;
         case 'deploy':
+          telemetry.trackCliCommandDeploy(userSuppliedSubCommand);
           func = require('./commands/deploy').default;
           break;
         case 'dev':
@@ -627,6 +629,7 @@ const main = async () => {
           func = require('./commands/remove').default;
           break;
         case 'rollback':
+          telemetry.trackCliCommandRollback(userSuppliedSubCommand);
           func = require('./commands/rollback').default;
           break;
         case 'target':
@@ -636,9 +639,11 @@ const main = async () => {
           func = require('./commands/teams').default;
           break;
         case 'telemetry':
+          telemetry.trackCliCommandTeams(userSuppliedSubCommand);
           func = require('./commands/telemetry').default;
           break;
         case 'whoami':
+          telemetry.trackCliCommandWhoami(userSuppliedSubCommand);
           func = require('./commands/whoami').default;
           break;
         default: