+++ packages/cli/src/util/telemetry/root.ts
@@ -8,6 +8,41 @@ export class RootTelemetryClient extends TelemetryClient {
     });
   }
 
+  trackCliCommandAlias(actual: string) {
+    this.trackCliCommand({
+      command: 'alias',
+      value: actual,
+    });
+  }
+
+  trackCliCommandDeploy(actual: string) {
+    this.trackCliCommand({
+      command: 'deploy',
+      value: actual,
+    });
+  }
+
+  trackCliCommandWhoami(actual: string) {
+    this.trackCliCommand({
+      command: 'whoami',
+      value: actual,
+    });
+  }
+
+  trackCliCommandRollback(actual: string) {
+    this.trackCliCommand({
+      command: 'rollback',
+      value: actual,
+    });
+  }
+
+  trackCliCommandTeams(actual: string) {
+    this.trackCliCommand({
+      command: 'teams',
+      value: actual,
+    });
+  }
+
   trackCPUs() {
     super.trackCPUs();
   }